requirements.txt
streamlit
opencv-python-headless
pytesseract
gspread
oauth2client
numpy
Pillow